# saas-forge-contracts/compatibility-baselines/v1/v0.1.0/openapi/v1.yaml
openapi: 3.1.1
info:
  title: saas-forge Public API
  version: 1.0.0
  description: |
    The versioned public REST contract for saas-forge. This document contains only
    delivery-stage-two and delivery-stage-three resources. Permission, Feature,
    and Quota runtime operations are intentionally absent until their domain
    contracts are reviewed.
servers:
  - url: https://api.example.invalid
    description: Deployment-specific API Gateway origin.
...
paths:
...
  /api/v1/platform/oauth-clients:
    post:
      tags: [OAuth clients]
      operationId: createOAuthClient
      x-saas.forge-service: iam-service
      summary: Create an OAuth Client and reveal its Secret once
      security: [{ UserBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOAuthClientRequest' }
      responses:
        '201':
          description: The OAuth Client was created. Store clientSecret now; it cannot be retrieved again.
          headers:
            Location: { $ref: './common.yaml#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthClientSecretResult' }
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
  /api/v1/platform/oauth-clients/{clientId}/secret-rotations:
    post:
      tags: [OAuth clients]
      operationId: rotateOAuthClientSecret
      x-saas.forge-service: iam-service
      summary: Rotate an OAuth Client Secret and reveal the new Secret once
      description: The previous and new Secret overlap for the service-controlled rotation window; revocation ends all Client access immediately.
      security: [{ UserBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: The Secret was rotated. Store clientSecret now; it cannot be retrieved again.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthClientSecretResult' }
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
        '404': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
  /api/v1/platform/oauth-clients/{clientId}/revocations:
    post:
      tags: [OAuth clients]
      operationId: revokeOAuthClient
      x-saas.forge-service: iam-service
      summary: Irreversibly revoke an OAuth Client
      security: [{ UserBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204': { $ref: './common.yaml#/components/responses/NoContent' }
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
        '404': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
components:
  securitySchemes:
    UserBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Platform Admin user Access Token. OAuth Client tokens cannot access platform management operations.
    OAuthClientBasic:
      type: http
      scheme: basic
      description: OAuth Client client_id and client_secret.
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { $ref: './common.yaml#/components/schemas/UuidV7' }
...
    ClientId:
      name: clientId
      in: path
      required: true
      schema: { $ref: './common.yaml#/components/schemas/UuidV7' }
  schemas:
    Timestamp:
      type: string
      format: date-time
      pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}[.][0-9]{3}Z$'
...
    RuntimeScope:
      type: string
      enum: ['runtime:read', 'runtime:quota:write']
    OAuthClientStatus:
      type: string
      enum: [ACTIVE, REVOKED]
...
    CreateOAuthClientRequest:
      type: object
      required: [displayName, allowedScopes]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 200 }
        allowedScopes:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/RuntimeScope' }
      additionalProperties: false
    OAuthClientSecretResult:
      type: object
      required: [clientId, displayName, allowedScopes, status, createdAt, updatedAt, clientSecret]
      properties:
        clientId: { $ref: './common.yaml#/components/schemas/UuidV7' }
        displayName: { type: string, minLength: 1, maxLength: 200 }
        allowedScopes:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/RuntimeScope' }
        status: { $ref: '#/components/schemas/OAuthClientStatus' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
        clientSecret:
          type: string
          minLength: 1
          writeOnly: true
          description: Returned exactly once. Store it in a controlled secret manager.
      additionalProperties: false